+++ out/pages/spriteeditor.js
@@ -86,6 +86,9 @@ class SpriteEditorPage {
                     case 'exportAsPngFile':
                         this.exportAsPngFile(message);
                         break;
+                    case 'exportAllAsPngFile':
+                        this.exportAllAsPngFile(message);
+                        break;
                     case 'exportAsBatariFile':
                         this.exportAsBatariFile(message);
                         break;
@@ -274,6 +277,68 @@ class SpriteEditorPage {
                             errorMessage: `Failed to export image '${path.basename(fileUri.fsPath)}' (Error: ${e.message})`
                         });
                     });
+                })
+                    .catch((e) => {
+                    this.currentPanel.webview.postMessage({
+                        command: command,
+                        status: 'error',
+                        errorMessage: e.message
+                    });
+                });
+            }
+        });
+    }
+    exportAllAsPngFile(message) {
+        // Prepare
+        let command = message.command;
+        let file = message.file;
+        let data = message.data;
+        // Get default path
+        // Assuiming file provided is the project file
+        let defaultUri = vscode.Uri.file(!file ? filesystem.WorkspaceFolder() : path.dirname(file));
+        // Prompt user here
+        let options = {
+            canSelectMany: false,
+            canSelectFolders: true,
+            canSelectFiles: false,
+            openLabel: "Export (Individual Files)",
+            defaultUri: defaultUri,
+            filters: {
+                'All Files': ['*']
+            }
+        };
+        // Process
+        vscode.window.showOpenDialog(options).then(folderUri => {
+            // Save?
+            if (folderUri) {
+                // Prepare
+                let folder = folderUri[0].fsPath;
+                // Save
+                filesystem.MkDirAsync(folder)
+                    .then(() => {
+                    // Process each image
+                    for (let i = 0; i < data.count; i++) {
+                        // Loop through each file
+                        let fileName = data.fileName + i + ".png";
+                        let fileUri = vscode.Uri.file(path.join(folder, fileName));
+                        // We now send through in base64 to avoid Buffer.from() conversion errors
+                        let convertToBuffer = Buffer.from(data.sprites[i], "base64");
+                        filesystem.WriteFileAsync(fileUri.fsPath, convertToBuffer)
+                            .then(() => {
+                            this.currentPanel.webview.postMessage({
+                                command: command,
+                                status: 'ok',
+                                file: path.basename(fileUri.fsPath),
+                            });
+                        })
+                            .catch((e) => {
+                            this.currentPanel.webview.postMessage({
+                                command: command,
+                                status: 'error',
+                                errorMessage: `Failed to export image '${path.basename(fileUri.fsPath)}' (Error: ${e.message})`
+                            });
+                        });
+                    }
                 })
                     .catch((e) => {
                     this.currentPanel.webview.postMessage({