+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The following enhancements and changes have been made to Atari Dev Studio:
 ## 0.6.4
 
 * Updated dasm to v2.20.14 (Windows, Linux, macOS)
+
+Sprite Editor
+* Added option to export all sprites to individual files
+
 ## 0.6.3
 
 * Updated 7800basic to v0.15 (Windows, Linux, macOS)